Pull from typeorm

+++ src/commands/CommandUtils.ts
@@ -17,10 +17,22 @@ export class CommandUtils {
     /**
      * Creates a file with the given content in the given path.
      */
-    static async createFile(filePath: string, content: string): Promise<void> {
+    static async createFile(filePath: string, content: string, override: boolean = true): Promise<void> {
         await CommandUtils.createDirectories(path.dirname(filePath));
         return new Promise<void>((ok, fail) => {
+            if (override === false && fs.existsSync(filePath))
+                return ok();
+
             fs.writeFile(filePath, content, err => err ? fail(err) : ok());
+        });
+    }
+
+    /**
+     * Reads everything from a given file and returns its content as a string.
+     */
+    static async readFile(filePath: string): Promise<string> {
+        return new Promise<string>((ok, fail) => {
+            fs.readFile(filePath, (err, data) => err ? fail(err) : ok(data.toString()));
         });
     }
